requirements.txt
flask
flask_sqlalchemy
flask_cors
flask_jwt_extended
flask_restful
flask_sse
celery
weasyprint
